server/requirements.txt
openenv-core
fastapi>=0.110.0
uvicorn>=0.29.0
pydantic>=2.0.0
datasets>=2.18.0
huggingface-hub>=0.21.0
openai>=1.0.0
requests>=2.31.0
python-dotenv>=1.0.0
